cli: Move the answer model out of agent/ into ai/, and parse every widget kind
`cli/agent/render.rs` still owned what an answer *is* — its segments, widget
references, inline markers, quote-card data — even though the `ai` TUI is its
main consumer and `agent/` is meant to hold the small standalone CLI commands.
Chart drawing moved to `ai/` last commit; this finishes the job.

`ai/answer.rs` now owns the model, and `cli/agent/render.rs` is 127 lines of
stdout presentation for `agent chat` alone. `strip_control_chars` is imported
from `utils::text` by its callers instead of being re-exported through a module
that no longer has anything to do with them.

Widget parsing was one function that understood one URL shape: a single
`symbol=` on a `widget://quote/` path. The web client's registry
(`portai/frontend/web/src/features/x-widget/index.tsx` `COMPONENTS_MAP`) lists
nine kinds, so everything else — a two-stock comparison, a stock list, an
open-account prompt, a backtest result — fell through and printed its URL into
the prose. There is now a `WidgetRef` for each: single quote, comparison, stock
list, CTA, and a catch-all that still carries its path, so a kind added
server-side is named rather than leaked as a URL.

Two things that shape the parser:

- A repeated key is how the agent passes a list (`symbols=A&symbols=B`), so
  values are collected rather than first-wins — otherwise a two-stock comparison
  quietly becomes one stock.
- `symbols=` must not be read as `symbol=`. The old prefix match was safe only by
  accident; both are now matched as whole keys.

A test caught the `#widget` fragment some URLs carry landing inside the last
parameter's value — `AAPL.US#widget` instead of `AAPL.US` — because the fragment
was being stripped from the path, which is the wrong end of the URL.

Rendering these kinds in the transcript is the next commit; this one is the model
and the move.
