+++ src/ai/mod.rs
@@ -2,6 +2,7 @@
 //!
 //! Structured after grok-build's layering, at a scale proportionate to a hosted
 //! chat agent:
+//! - [`answer`]  — what an answer is made of (segments, widgets, markers)
 //! - [`chart`]   — `vis-chart` specs drawn as braille plots
 //! - [`state`]   — the chat state snapshot + event model (grok's `xai-chat-state`)
 //! - [`runtime`] — the agent-runtime seam that streams a turn (grok's `xai-grok-shell`)
@@ -11,6 +12,7 @@
 //! this reuses the shared streaming in [`crate::cli::agent::client`] and has no
 //! local tool/workspace layer (unlike grok-build, which edits and runs code).
 
+pub mod answer;
 pub mod chart;
 pub mod editor;
 pub mod markdown;
